+++ packages/flutter_preact_signals/lib/flutter_preact_signals.dart
@@ -1,4 +1,7 @@
 library flutter_preact_signals;
 
 export 'package:preact_signals/preact_signals.dart';
-export 'src/extensions.dart';
+export 'src/extensions/signal.dart';
+export 'src/extensions/value_listenable.dart';
+export 'src/extensions/value_notifier.dart';
+export 'src/watch_signal.dart';