+++ packages/flutter_preact_signals/lib/src/extensions.dart
@@ -1,2 +1,0 @@
-export 'extensions/notifier.dart';
-export 'extensions/watch.dart';