+++ packages/flutter_preact_signals/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.1
+
+- Fixing `watchSignal` so it works in StatefulWidgets and inside rebuilds and subscriber list will remain stable
+- Updating example to show non global signal
+- Adding extension methods on `ReadonlySignal`: `stream`, `listenable`
+- Adding extension methods on `MutableSignal`: `notifier`, `readonly`
+- Updating `preact_signals` to 0.3.0
+
 ## 0.2.0
 
 - Updating `preact_signals` to 0.2.1
@@ -14,7 +22,7 @@
 ## 0.0.2
 
 - Fixing recursive call after refactor
-- 
+
 ## 0.0.1
 
 - Adding extensions for `ValueNotifier`, `ValueListenable` and `BuildContext`